fix failed ut
