--- python/chronos/test/bigdl/chronos/autots/model/utils.py
@@ -1,43 +1,0 @@
-#
-# Copyright 2016 The BigDL Authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-from bigdl.chronos.utils import LazyImport
-torch = LazyImport('torch')
-Dataset = LazyImport('torch.utils.data')
-DataLoader = LazyImport('torch.utils.data')
-import numpy as np
-
-input_feature_dim = 10
-output_feature_dim = 2
-past_seq_len = 5
-future_seq_len = 1
-
-def get_x_y(size):
-    x = np.random.randn(size, past_seq_len, input_feature_dim)
-    y = np.random.randn(size, future_seq_len, output_feature_dim)
-    return x.astype(np.float32), y.astype(np.float32)
-
-class RandomDataset(Dataset):
-    def __init__(self, size=1000):
-        x, y = get_x_y(size)
-        self.x = torch.from_numpy(x).float()
-        self.y = torch.from_numpy(y).float()
-
-    def __len__(self):
-        return self.x.shape[0]
-
-    def __getitem__(self, idx):
-        return self.x[idx], self.y[idx]
